Отработка на Windows

+++ errorRate.cpp
@@ -7,7 +7,7 @@ std::istream& operator>>(std::istream& is, Frames& fr) {
   return is;
 }
 
-float getErrorRateIntervals(const Frames& last, const Frames& next) {
+double getErrorRateIntervals(const Frames& last, const Frames& next) {
   return (next.lost_frames - last.lost_frames) / ((next.frames - last.frames) * 1.0);
 }
 
@@ -15,7 +15,7 @@ void ErrorRate::setCurrent(uint64_t time, const Frames& frames) {
   mErrorRate[time] = frames;
 }
 
-float ErrorRate::getErrorRate(uint32_t duration) const {
+double ErrorRate::getErrorRate(uint32_t duration) const {
   if (mErrorRate.empty()) {
     return 0;
   }
@@ -26,8 +26,8 @@ float ErrorRate::getErrorRate(uint32_t duration) const {
   if (mErrorRate.rbegin()->first > duration) {
     it = mErrorRate.lower_bound(mErrorRate.rbegin()->first - duration);
   }
-  float sumER = 0;
-  float sumDurationIntervals = 0;
+  double sumER = 0;
+  double sumDurationIntervals = 0;
   while (next(it) != mErrorRate.end()) {
     uint64_t durationIntervals = next(it)->first - it->first;
     sumER += getErrorRateIntervals(it->second, next(it)->second) * durationIntervals;
